# Backend/src/main/webapp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ISDCM Video API",
    "version": "1.0",
    "description": "API RESTful para la gestión de vídeos en la plataforma ISDCM."
  },
  "tags": [
    {
      "name": "Autenticación",
      "description": "Operaciones de autenticación de usuarios"
    },
    {
      "name": "Videos",
      "description": "Operaciones relacionadas con la gestión de vídeos"
    }
  ],
  "servers": [
    {
      "url": "/Backend/resources",
      "description": "API base path"
    }
  ],
  "paths": {
    "/videos": {
      "get": {
        "summary": "Listar todos los vídeos",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "top",
            "in": "query",
            "description": "Número máximo de resultados a devolver",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de vídeos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Registrar un nuevo vídeo",
        "tags": [
          "Videos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vídeo creado"
          },
          "500": {
            "description": "Error al registrar el vídeo"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/videos/buscar": {
      "get": {
        "summary": "Buscar vídeos con filtros",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "titulo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "autor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fecha",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "top",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "orden",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "fecha",
                "vistas"
              ]
            }
          },
          {
            "name": "dir",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de la búsqueda",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/videos/{id}": {
      "get": {
        "summary": "Obtener vídeo por ID",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vídeo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDTO"
                }
              }
            }
          },
          "404": {
            "description": "Vídeo no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Actualizar vídeo por ID",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vídeo actualizado"
          },
          "404": {
            "description": "Vídeo no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Eliminar vídeo por ID",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vídeo eliminado"
          },
          "404": {
            "description": "Vídeo no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/videos/visualizar/{id}": {
      "post": {
        "summary": "Incrementar contador de visualización",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reproducción contada"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/videos/{id}/propietario": {
      "get": {
        "summary": "Verificar si un usuario es propietario del vídeo",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado de la verificación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/videos/existe": {
      "get": {
        "summary": "Verificar si un vídeo ya fue registrado por un usuario",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "titulo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Boolean indicando existencia",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/login": {
      "post": {
        "summary": "Iniciar sesión y obtener JWT",
        "description": "Permite a un usuario autenticarse y recibir un JWT para futuras solicitudes.",
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password" 
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT generado con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "JWT": {
                      "type": "string",
                      "description": "Token JWT generado"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "VideoDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "titulo": {
            "type": "string"
          },
          "autor": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "duracion": {
            "type": "string"
          },
          "formato": {
            "type": "string"
          },
          "fechaCreacion": {
            "type": "string"
          },
          "localizacion": {
            "type": "string"
          },
          "userId": {
            "type": "integer"
          },
          "numReproducciones": {
            "type": "integer"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}